+++ docs/2019-09-15_OfficialDocumentation/class_io_taa_p___accelerometer.js
@@ -1,0 +1,6 @@
+var class_io_taa_p___accelerometer =
+[
+    [ "IoTaaP_Accelerometer", "class_io_taa_p___accelerometer.html#a4ba847f2f4cfdfa261496dd74e451f43", null ],
+    [ "getRaw", "class_io_taa_p___accelerometer.html#a0a50a823b4cd3204c93824d3bca19394", null ],
+    [ "adc", "class_io_taa_p___accelerometer.html#a82ec12033fb00da80ece85d89b9b41ce", null ]
+];